moved SelectionDropdown into DropdownMenu.tsx
